1. Get current Location from IP 2. Set Temperature as Integer instead of Decimal 3. arg checking update to ignore empty value for Provider or Location

+++ nahid/xWeather/Drivers/WeatherApi/WeatherApiDriver.cs
@@ -34,7 +34,7 @@ namespace xWeather.Drivers.WeatherApi
                     {
                         Location = location,
                         Condition = weatherResponse.Current.Condition.Text,
-                        Temperature = weatherResponse.Current.Temperature,
+                        Temperature = (int) weatherResponse.Current.Temperature,
                         Humidity = weatherResponse.Current.Humidity,
                         WindSpeed = weatherResponse.Current.WindSpeed,
                         Pressure = weatherResponse.Current.Pressure,
